requirements.txt
arrow==0.14.2
Click==7.0
dnspython==1.16.0
eventlet==0.25.0
greenlet==0.4.15
monotonic==1.5
pkg-resources==0.0.0
pony==0.7.10
python-dateutil==2.8.0
PyYAML==5.1.1
six==1.12.0
tqdm==4.32.1